requirements.txt
# Mac / PC 側クライアントの最小要件
requests>=2.31.0
numpy>=1.25
opencv-python>=4.8
pillow>=10.0
num2words>=0.5
mlx>=0.12.0; sys_platform == "darwin"
mlx-vlm>=0.1.0; sys_platform == "darwin"
